Working multiupload
We return a list of results and if there is more files we append those results to that list.

Attempted to use uploader.js and jquery multiupload but it seems to be more complicated than what I need.  Drag and drop would be nice but I would rather it worked first.

Signed-off-by: binkybear <binkybear@nethunter.com>